Remove UDP related code from JavaScript
