+++ js/src/Ice/EndpointInfo.d.ts
@@ -11,7 +11,7 @@ export namespace IceSSL
     {
         /**
          * One-shot constructor to initialize all data members.
-         * @param underlying The information of the underyling endpoint or null if there's no underlying endpoint.
+         * @param underlying The information of the underlying endpoint or null if there's no underlying endpoint.
          * @param timeout The timeout for the endpoint in milliseconds.
          * @param compress Specifies whether or not compression should be used if available when using this endpoint.
          */
