+++ js/src/Ice/ImplicitContext.js
@@ -2,25 +2,147 @@
 // Copyright (c) ZeroC, Inc. All rights reserved.
 //
 
-/* eslint-disable */
-/* jshint ignore: start */
+const Ice = require("../Ice/ModuleRegistry").Ice;
 
-/* slice2js browser-bundle-skip */
-const _ModuleRegistry = require("../Ice/ModuleRegistry").Ice._ModuleRegistry;
-require("../Ice/Object");
-require("../Ice/Value");
-require("../Ice/ObjectPrx");
-require("../Ice/Long");
-require("../Ice/HashMap");
-require("../Ice/HashUtil");
-require("../Ice/ArrayUtil");
-require("../Ice/StreamHelpers");
-require("../Ice/LocalException");
 require("../Ice/Current");
-const Ice = _ModuleRegistry.module("Ice");
+require("../Ice/LocalException");
 
-const Slice = Ice.Slice;
-/* slice2js browser-bundle-skip-end */
-/* slice2js browser-bundle-skip */
-exports.Ice = Ice;
-/* slice2js browser-bundle-skip-end */
+const Context = Ice.Context;
+const InitializationException = Ice.InitializationException;
+
+//
+// The base class for all ImplicitContext implementations
+//
+class ImplicitContext
+{
+    constructor()
+    {
+        this._context = new Context();
+    }
+
+    getContext()
+    {
+        return new Context(this._context);
+    }
+
+    setContext(context)
+    {
+        if(context !== null && context.size > 0)
+        {
+            this._context = new Context(context);
+        }
+        else
+        {
+            this._context.clear();
+        }
+    }
+
+    containsKey(key)
+    {
+        if(key === null)
+        {
+            key = "";
+        }
+
+        return this._context.has(key);
+    }
+
+    get(key)
+    {
+        if(key === null)
+        {
+            key = "";
+        }
+
+        let val = this._context.get(key);
+        if(val === null)
+        {
+            val = "";
+        }
+
+        return val;
+    }
+
+    put(key, value)
+    {
+        if(key === null)
+        {
+            key = "";
+        }
+        if(value === null)
+        {
+            value = "";
+        }
+
+        let oldVal = this._context.get(key);
+        if(oldVal === null)
+        {
+            oldVal = "";
+        }
+
+        this._context.set(key, value);
+
+        return oldVal;
+    }
+
+    remove(key)
+    {
+        if(key === null)
+        {
+            key = "";
+        }
+
+        let val = this._context.get(key);
+        this._context.delete(key);
+
+        if(val === null)
+        {
+            val = "";
+        }
+        return val;
+    }
+
+    write(prxContext, os)
+    {
+        if(prxContext.size === 0)
+        {
+            Ice.ContextHelper.write(os, this._context);
+        }
+        else
+        {
+            let ctx = null;
+            if(this._context.size === 0)
+            {
+                ctx = prxContext;
+            }
+            else
+            {
+                ctx = new Context(this._context);
+                for(const [key, value] of prxContext)
+                {
+                    ctx.set(key, value);
+                }
+            }
+            Ice.ContextHelper.write(os, ctx);
+        }
+    }
+
+    static create(kind)
+    {
+        if(kind.length === 0 || kind === "None")
+        {
+            return null;
+        }
+        else if(kind === "Shared")
+        {
+            return new ImplicitContext();
+        }
+        else
+        {
+            throw new InitializationException("'" + kind + "' is not a valid value for Ice.ImplicitContext");
+        }
+    }
+}
+
+Ice.ImplicitContext = ImplicitContext;
+module.exports.Ice = Ice;
