+++ js/src/Ice/sources.json
@@ -52,7 +52,6 @@
         "IPEndpointI.js",
         "IdentityUtil.js",
         "ImplicitContext.js",
-        "ImplicitContextI.js",
         "IncomingAsync.js",
         "Initialize.js",
         "Instance.js",
@@ -61,7 +60,6 @@
         "LocatorManager.js",
         "LocatorTable.js",
         "Logger.js",
-        "LoggerI.js",
         "Long.js",
         "MapUtil.js",
         "Object.js",
